🐛 fix(ui): handle error response in dialogs ✨ feat(ui): adjust font size based on text length ✨ feat(ui): add support for formdata body type in thunderclient.json The changes in the UI files add support for handling error responses in dialogs. The font size in the SpeakingExerciseTopSection component is now adjusted based on the length of the Chinese text. The thunderclient.json file now supports formdata as a body type.

+++ lib/ui/pages/profile/detail_profile_ui.dart
@@ -125,28 +125,38 @@ class _ProfileDetailPageState extends State<ProfileDetailPage> {
           },
           loaded: (user) async {
             if (context.mounted) {
-              SuccessDialog(
-                context: context,
-                desc:
-                    'Profile has been updated, for the changes to take effect, we`ll need to log you out.',
-                title: 'Success',
-                btnOkOnPress: () {
-                  AppSecureStorage.deleteAll().then(
-                    (value) {
-                      context.read<UserCubit>().clearUser();
+              if (user.status == 'Error') {
+                ErrorDialog(
+                  context: context,
+                  title: 'Error',
+                  desc: user.message!,
+                  btnOkOnPress: () {},
+                  btnOkText: 'OK',
+                ).show();
+              } else {
+                SuccessDialog(
+                  context: context,
+                  desc:
+                      'Profile has been updated, for the changes to take effect, we`ll need to log you out.',
+                  title: 'Success',
+                  btnOkOnPress: () {
+                    AppSecureStorage.deleteAll().then(
+                      (value) {
+                        context.read<UserCubit>().clearUser();
 
-                      GoogleSignIn().disconnect();
+                        GoogleSignIn().disconnect();
 
-                      Navigator.pushNamedAndRemoveUntil(
-                        context,
-                        LoginPage.routeName,
-                        (route) => false,
-                      );
-                    },
-                  );
-                },
-                btnOkText: 'OK',
-              ).show();
+                        Navigator.pushNamedAndRemoveUntil(
+                          context,
+                          LoginPage.routeName,
+                          (route) => false,
+                        );
+                      },
+                    );
+                  },
+                  btnOkText: 'OK',
+                ).show();
+              }
             }
           },
         );
